delete fonctionnel, manque favori, modification MDP et le css

+++ Controllers/AdController.js
@@ -3,6 +3,18 @@ const db = require('../db/db');
 const adsView = require('../Views/adsView');
 const createAdView = require('../Views/createAdView');
 const pendingAdsView = require('../Views/pendingAdsView');
+const userAdsView = require('../Views/userAdsView');
+const allAdsView = require('../Views/allAdsView');
+
+function redirectScript(url, delay) {
+    return `
+        <script>
+            setTimeout(function() {
+                window.location.href = '${url}';
+            }, ${delay});
+        </script>
+    `;
+}
 
 function showAds(req, res) {
     const query = `
@@ -35,12 +47,8 @@ function createAd(req, res) {
         } else {
             res.send(`
                 <p>Ad created successfully, pending approval</p>
-                <p>You will be redirected to the ads page in 5 seconds...</p>
-                <script>
-                    setTimeout(function() {
-                        window.location.href = '/ads';
-                    }, 5000);
-                </script>
+                <p>You will be redirected to the ads page in 1 seconds...</p>
+                ${redirectScript('/ads', 1000)}
             `);
         }
     });
@@ -54,16 +62,12 @@ function approveAd(req, res) {
             console.error('Error approving ad: ', err.message);
             res.send('Error approving ad');
         } else {
-        res.send(`
-            <p>Ad created successfully</p>
-            <p>You will be redirected to the ads page in 5 seconds...</p>
-            <script>
-                setTimeout(function() {
-                    window.location.href = '/admin';
-                }, 5000);
-            </script>
-        `);
-    }
+            res.send(`
+                <p>Ad approved successfully</p>
+                <p>You will be redirected to the admin panel page in 1 seconds...</p>
+                ${redirectScript('/admin', 1000)}
+            `);
+        }
     });
 }
 
@@ -77,12 +81,8 @@ function disapproveAd(req, res) {
         } else {
             res.send(`
                 <p>Ad deleted successfully</p>
-                <p>You will be redirected to the ads page in 5 seconds...</p>
-                <script>
-                    setTimeout(function() {
-                        window.location.href = '/admin';
-                    }, 5000);
-                </script>
+                <p>You will be redirected to the admin panel page in 1 seconds...</p>
+                ${redirectScript('/admin', 1000)}
             `);
         }
     });
@@ -90,17 +90,35 @@ function disapproveAd(req, res) {
 
 function deleteAd(req, res) {
     const adId = req.params.id;
-    const query = `DELETE FROM ads WHERE id = ?`;
-    db.run(query, [adId], function (err) {
+    const userId = req.user.id;
+    const isAdmin = req.user.isAdmin;
+
+    console.log(`Attempting to delete ad with ID: ${adId} by user ID: ${userId} (Admin: ${isAdmin})`);
+
+    const query = isAdmin ? `DELETE FROM ads WHERE id = ?` : `DELETE FROM ads WHERE id = ? AND userId = ?`;
+    const params = isAdmin ? [adId] : [adId, userId];
+
+    db.run(query, params, function (err) {
         if (err) {
             console.error('Error deleting ad: ', err.message);
             res.send('Error deleting ad');
         } else {
-            res.send('Ad deleted successfully');
+            if (this.changes > 0) {
+                res.send(`
+                    <p>Ad deleted successfully</p>
+                    <p>You will be redirected in 1 second...</p>
+                    ${redirectScript(isAdmin ? '/admin' : '/ads/user', 1000)}
+                `);
+            } else {
+                res.send(`
+                    <p>Ad not found or you do not have permission to delete it</p>
+                    <p>You will be redirected in 3 seconds...</p>
+                    ${redirectScript(isAdmin ? '/admin' : '/ads/user', 1000)}
+                `);
+            }
         }
     });
 }
-
 
 function showPendingAds(req, res) {
     const query = `SELECT * FROM ads WHERE isApproved = 0`;
@@ -114,4 +132,38 @@ function showPendingAds(req, res) {
     });
 }
 
-module.exports = { showAds, showCreateAd, createAd, approveAd, deleteAd, showPendingAds, disapproveAd };
+function showUserAds(req, res) {
+    const userId = req.user.id;
+    const query = `
+        SELECT ads.*, users.username 
+        FROM ads 
+        JOIN users ON ads.userId = users.id 
+        WHERE ads.userId = ?
+    `;
+    db.all(query, [userId], (err, rows) => {
+        if (err) {
+            console.error('Error fetching user ads: ', err.message);
+            res.send('Error loading user ads');
+        } else {
+            res.send(userAdsView(rows, req.user));
+        }
+    });
+}
+
+function showAllAds(req, res) {
+    const query = `
+        SELECT ads.*, users.username 
+        FROM ads 
+        JOIN users ON ads.userId = users.id
+    `;
+    db.all(query, [], (err, rows) => {
+        if (err) {
+            console.error('Error fetching all ads: ', err.message);
+            res.send('Error loading all ads');
+        } else {
+            res.send(allAdsView(rows, req.user));
+        }
+    });
+}
+
+module.exports = { showAds, showCreateAd, createAd, approveAd, deleteAd, showPendingAds, disapproveAd, showUserAds, showAllAds };
